[cuda][fix] average pool float to half

+++ src/ppl/nn/engines/cuda/impls/src/nn/global_pooling_ave.cu
@@ -125,7 +125,7 @@ __global__ void ppl_cukernel_pooling_ave_global_shuffle_half(
 
     // store output
     if (threadIdx.x == 0)
-        output[bc] = __half2float(res / HW);
+        output[bc] = __float2half(res / HW);
 #endif
 }
